Nevin x

+++ app/src/main/java/com/example/eventsnapqr/FirebaseController.java
@@ -170,9 +170,40 @@ public class FirebaseController {
      * @param userId
      */
     private void deleteUserFinalStep(FirebaseFirestore db, String userId) {
-        db.collection("users").document(userId).delete()
-                .addOnSuccessListener(aVoid -> Log.d("Delete User", "User successfully deleted: " + userId))
-                .addOnFailureListener(e -> Log.e("Delete User", "Error deleting user: " + userId, e));
+        CollectionReference notificationReference = db.collection("users").document(userId).collection("notifications");
+        notificationReference.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+            @Override
+            public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                if (task.isSuccessful()) {
+                    QuerySnapshot snapshot = task.getResult();
+                    if (snapshot.size() > 0) {
+                        int[] i = {0};
+                        for (QueryDocumentSnapshot doc : snapshot) {
+                            notificationReference.document(doc.getId()).delete().addOnCompleteListener(new OnCompleteListener<Void>() {
+                                @Override
+                                public void onComplete(@NonNull Task<Void> task) {
+                                    if (i[0] == snapshot.size() - 1) {
+                                        db.collection("users").document(userId).delete()
+                                                .addOnSuccessListener(aVoid -> Log.d("Delete User", "User successfully deleted: " + userId))
+                                                .addOnFailureListener(e -> Log.e("Delete User", "Error deleting user: " + userId, e));
+                                    }
+                                    i[0]++;
+                                }
+                            });
+                        }
+                    }
+                    else {
+                        db.collection("users").document(userId).delete()
+                                .addOnSuccessListener(aVoid -> Log.d("Delete User", "User successfully deleted: " + userId))
+                                .addOnFailureListener(e -> Log.e("Delete User", "Error deleting user: " + userId, e));
+                    }
+                } else {
+                    db.collection("users").document(userId).delete()
+                            .addOnSuccessListener(aVoid -> Log.d("Delete User", "User successfully deleted: " + userId))
+                            .addOnFailureListener(e -> Log.e("Delete User", "Error deleting user: " + userId, e));
+                }
+            }
+        });
     }
 
     private Task<Void> fetchAndDeleteEvent(FirebaseFirestore db, String eventId) {
